refactor(docs): calendar dx
