+++ apps/docs/content/components/calendar/controlled.ts
@@ -1,39 +1,15 @@
-const App = `import {Calendar} from "@nextui-org/react";
-import {parseDate} from "@internationalized/date";
-
-export default function App() {
-  let [value, setValue] = React.useState(parseDate("2024-03-07"));
-
-  return (
-    <Calendar 
-      aria-label="Date (Controlled)" 
-      value={value} 
-      onChange={setValue} 
-    />
-  );
-}`;
-
-const AppTs = `import {Calendar} from "@nextui-org/react";
-import type {DateValue} from "@react-types/calendar";
-import {parseDate} from "@internationalized/date";
-
-export default function App() {
-  let [value, setValue] = React.useState<DateValue>(parseDate("2024-03-07"));
-
-  return (
-    <Calendar 
-      aria-label="Date (Controlled)" 
-      value={value} 
-      onChange={setValue} 
-    />
-  );
-}`;
+import App from "./controlled.raw.jsx?raw";
+import AppTs from "./controlled.raw.tsx?raw";
 
 const react = {
   "/App.jsx": App,
+};
+
+const reactTs = {
   "/App.tsx": AppTs,
 };
 
 export default {
   ...react,
+  ...reactTs,
 };
