+++ apps/docs/content/components/calendar/disabled.raw.jsx
@@ -1,0 +1,5 @@
+import {Calendar} from "@nextui-org/react";
+
+export default function App() {
+  return <Calendar isDisabled aria-label="Date (Disabled)" />;
+}
