+++ apps/docs/content/components/calendar/international-calendars.ts
@@ -1,13 +1,4 @@
-const App = `import {Calendar} from "@nextui-org/react";
-import {I18nProvider} from "@react-aria/i18n";
-
-export default function App() {
-  return (
-    <I18nProvider locale="zh-CN-u-ca-chinese">
-      <Calendar aria-label="Date (International Calendar)" />
-    </I18nProvider>
-  );
-}`;
+import App from "./international-calendars.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
