+++ apps/docs/content/components/calendar/invalid-date.ts
@@ -1,51 +1,15 @@
-const App = `import {Calendar} from "@nextui-org/react";
-import {today, getLocalTimeZone, isWeekend} from "@internationalized/date";
-import {useLocale} from "@react-aria/i18n";
-
-
-export default function App() {
-  let [date, setDate] = React.useState(today(getLocalTimeZone()));
-  let {locale} = useLocale();
-  let isInvalid = isWeekend(date, locale);
-
-  return (
-    <Calendar
-      aria-label="Date (Invalid on weekends)"
-      errorMessage={isInvalid ? "We are closed on weekends" : undefined}
-      isInvalid={isInvalid}
-      value={date}
-      onChange={setDate}
-    />
-  );
-}`;
-
-const AppTs = `import {Calendar} from "@nextui-org/react";
-import type {DateValue} from "@react-types/calendar";
-import {today, getLocalTimeZone, isWeekend} from "@internationalized/date";
-import {useLocale} from "@react-aria/i18n";
-
-
-export default function App() {
-  let [date, setDate] = React.useState<DateValue>(today(getLocalTimeZone()));
-  let {locale} = useLocale();
-  let isInvalid = isWeekend(date, locale);
-
-  return (
-    <Calendar
-      aria-label="Date (Invalid on weekends)"
-      errorMessage={isInvalid ? "We are closed on weekends" : undefined}
-      isInvalid={isInvalid}
-      value={date}
-      onChange={setDate}
-    />
-  );
-}`;
+import App from "./invalid-date.raw.jsx?raw";
+import AppTs from "./invalid-date.raw.tsx?raw";
 
 const react = {
   "/App.jsx": App,
+};
+
+const reactTs = {
   "/App.tsx": AppTs,
 };
 
 export default {
   ...react,
+  ...reactTs,
 };
