+++ apps/docs/content/components/calendar/max-date-value.ts
@@ -1,15 +1,4 @@
-const App = `import {Calendar} from "@nextui-org/react";
-import {today, getLocalTimeZone} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <Calendar
-      aria-label="Date (Max Date Value)"
-      defaultValue={today(getLocalTimeZone())}
-      maxValue={today(getLocalTimeZone())}
-    />
-  );
-}`;
+import App from "./max-date-value.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
