+++ apps/docs/content/components/calendar/page-behaviour.ts
@@ -1,13 +1,4 @@
-const App = `import {Calendar} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Calendar 
-      aria-label="Date (Page Behaviour)" 
-      pageBehavior="single" 
-    />
-  );
-}`;
+import App from "./page-behaviour.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
