+++ apps/docs/content/components/calendar/read-only.raw.jsx
@@ -1,0 +1,6 @@
+import {Calendar} from "@nextui-org/react";
+import {today, getLocalTimeZone} from "@internationalized/date";
+
+export default function App() {
+  return <Calendar isReadOnly aria-label="Date (Read Only)" value={today(getLocalTimeZone())} />;
+}
