+++ apps/docs/content/components/calendar/read-only.ts
@@ -1,15 +1,4 @@
-const App = `import {Calendar} from "@nextui-org/react";
-import {today, getLocalTimeZone} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <Calendar 
-      aria-label="Date (Read Only)" 
-      value={today(getLocalTimeZone())} 
-      isReadOnly 
-    />
-  );
-}`;
+import App from "./read-only.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
