+++ apps/docs/content/components/calendar/usage.ts
@@ -1,14 +1,4 @@
-const App = `import {Calendar} from "@nextui-org/react";
-import {parseDate} from '@internationalized/date';
-
-export default function App() {
-  return (
-    <div className="flex gap-x-4">
-      <Calendar aria-label="Date (No Selection)" />
-      <Calendar aria-label="Date (Uncontrolled)" defaultValue={parseDate("2020-02-03")} />
-    </div>
-  );
-}`;
+import App from "./usage.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
