+++ apps/docs/content/components/calendar/visible-months.raw.jsx
@@ -1,0 +1,5 @@
+import {Calendar} from "@nextui-org/react";
+
+export default function App() {
+  return <Calendar aria-label="Date (Visible Month)" visibleMonths={3} />;
+}
