+++ apps/docs/content/components/calendar/visible-months.ts
@@ -1,13 +1,4 @@
-const App = `import {Calendar} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Calendar 
-      aria-label="Date (Visible Month)" 
-      visibleMonths={3} 
-    />
-  );
-}`;
+import App from "./visible-months.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
