+++ apps/docs/content/components/calendar/with-month-and-year-picker.raw.jsx
@@ -1,0 +1,5 @@
+import {Calendar} from "@nextui-org/react";
+
+export default function App() {
+  return <Calendar showMonthAndYearPickers aria-label="Date (Show Month and Year Picker)" />;
+}
